requirements.txt
ultralytics==8.3.70
opencv-python-headless==4.9.0.80
opencv-contrib-python-headless==4.9.0.80
numpy==1.26.4
flask==3.0.2
gunicorn==21.2.0
requests==2.31.0
PyYAML==6.0.1
lapx>=0.5.5
filterpy==1.4.5
scikit-learn==1.4.1.post1
tqdm==4.66.2
python-telegram-bot==21.0.1
sqlalchemy==2.0.28
boto3==1.34.51
shapely==2.0.3
imgaug==0.4.0
albumentations==1.4.0
schedule==1.2.1
psutil==5.9.8
flake8==7.0.0
black==24.2.0
pytest==8.0.2
kaggle==1.6.6
